upgrade to v1.6

+++ src/Sylius/Component/Core/Customer/Statistics/CustomerStatistics.php
@@ -15,14 +15,9 @@ namespace Sylius\Component\Core\Customer\Statistics;
 
 use Webmozart\Assert\Assert;
 
-/**
- * @author Jan Góralski <jan.goralski@lakion.com>
- */
 final class CustomerStatistics
 {
-    /**
-     * @var array|PerChannelCustomerStatistics[]
-     */
+    /** @var array|PerChannelCustomerStatistics[] */
     private $perChannelsStatistics;
 
     /**
@@ -35,12 +30,9 @@ final class CustomerStatistics
         $this->perChannelsStatistics = $perChannelStatistics;
     }
 
-    /**
-     * @return int
-     */
     public function getAllOrdersCount(): int
     {
-        return array_sum(array_map(function (PerChannelCustomerStatistics $statistics) {
+        return (int) array_sum(array_map(function (PerChannelCustomerStatistics $statistics) {
             return $statistics->getOrdersCount();
         }, $this->perChannelsStatistics));
     }
