+++ src/Sylius/Component/Core/Customer/Statistics/CustomerStatisticsProvider.php
@@ -19,25 +19,14 @@ use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Repository\OrderRepositoryInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 
-/**
- * @author Jan Góralski <jan.goralski@lakion.com>
- */
 final class CustomerStatisticsProvider implements CustomerStatisticsProviderInterface
 {
-    /**
-     * @var OrderRepositoryInterface
-     */
+    /** @var OrderRepositoryInterface */
     private $orderRepository;
 
-    /**
-     * @var RepositoryInterface
-     */
+    /** @var RepositoryInterface */
     private $channelRepository;
 
-    /**
-     * @param OrderRepositoryInterface $orderRepository
-     * @param RepositoryInterface $channelRepository
-     */
     public function __construct(OrderRepositoryInterface $orderRepository, RepositoryInterface $channelRepository)
     {
         $this->orderRepository = $orderRepository;
@@ -75,12 +64,10 @@ final class CustomerStatisticsProvider implements CustomerStatisticsProviderInte
 
     /**
      * @param array|OrderInterface[] $orders
-     *
-     * @return int
      */
     private function getOrdersSummedTotal(array $orders): int
     {
-        return array_sum(
+        return (int) array_sum(
             array_map(function (OrderInterface $order) {
                 return $order->getTotal();
             }, $orders)
@@ -89,7 +76,6 @@ final class CustomerStatisticsProvider implements CustomerStatisticsProviderInte
 
     /**
      * @param array|OrderInterface[] $orders
-     * @param ChannelInterface $channel
      *
      * @return array|OrderInterface[]
      */
